Deleted CHANGELOG line not belonging to patch 2
Note: This is an exception to the rule that 'tags' sub-directories should not
be edited. It is too convenient not to do it in this case, because we are
trying to re-create a historic state of the repository for the sake of being
able to check it out with a symbolic tag name later.