New feature for modpatch: If $AUTO_FIX_PATCHES == "y", then the applied patch is analysed and automatically fixed, if fuzzy. Now what the h... does that mean and how does it work?   1. Perform dry run in order to check if the patch will be both successful      *and* fuzzy (i.e. containing moved or fuzzy-fitting hunks).   2. If so, enter auto-fixing mode and generate a list of all files changed by      the patch.   3. Apply patch with the option to create a 'foo.orig' backup for each      patched file 'foo'.   4. Finally, auto-fix the original fuzzy patch by creating a new one, cycling      through each pair 'foo.orig' / 'foo', creating a fresh 'diff' for them,      but preserving the old patch as 'xy.patch.orig'.   5. The 'foo.orig' files are cleaned up once they are not needed anymore, but      the 'xy.patch.orig' files are being kept so they can be compared to their      auto-fixed versions. Nobody is perfect, so a closer look should be taken.
The next time 'fwmod' is run with the newly created, polished-up patches, they
should all fit perfectly - no moved hunks, no fuzzy matches. As an exception,
modpatch does not try to fix patches containing failed hunks but exits 'fwmod'
as usual in this case, because the patch needs to be fixed anyway. It should
be enough to fix it roughly so it can be applied as a fuzzy patch - modpatch
can do the rest in the next run. ;-)