Introduce awk script + shell helper script in order to copy with JavaScript function definitions to be removed in patches/220-help.sh and containing nested curly braces.
This makes "remove help" rather SLOW, and I still think that "sed patches" are
a nightmare to maintain and test, but at least it gave me an opportunity to
learn a little bit of awk language.

Still to do: Some HTML and JavaScript files still refer to the functions
removed (uiDoHelp*, jslPopHelp*). I did not touch that part for now, it is
so boring.