New target `config-compress` calls Kconfig's `--savedefconfig` option in order to create a minimised ("compressed") version of ''.config'' showing only user-selected, non-default options. This is a great debug helper to quickly determine what is "special" about a certain configuration.
BTW: It is even possible to rename ''.config_compressed'' to ''.config'' and
re-save it via ''menuconfig'' or ''oldnoconfig'', i.e. the two versions are
really equivalent.

I also added a comment to macro `CONFIG_CLEAN_DEPS` which explains that targets
like `config-clean-deps` currently do not deactivate all deactivateable menu
options for the selected categories, but just those with a default of "n".
Those with default "y/m" will remain active. I guess not everybody knew that.